# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LeaderboardEntry {
  id: string;
  playerName: string;
  wave: number;
  kills: number;
  scrap: number;
  createdAt: string;
}

export interface LeaderboardInput {
  /**
     * @minLength 1
     * @maxLength 24
     */
  playerName: string;
  /** @minimum 1 */
  wave: number;
  /** @minimum 0 */
  kills: number;
  /** @minimum 0 */
  scrap: number;
}

export interface LeaderboardSubmitResult {
  ranked: boolean;
  /** @nullable */
  rank: number | null;
}