# lib/api-zod/src/generated/types/leaderboardInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardInput {
  /**
     * @minLength 1
     * @maxLength 24
     */
  playerName: string;
  /** @minimum 1 */
  wave: number;
  /** @minimum 0 */
  kills: number;
  /** @minimum 0 */
  scrap: number;
}