# lib/api-zod/src/generated/types/leaderboardSubmitResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardSubmitResult {
  ranked: boolean;
  /** @nullable */
  rank: number | null;
}